Move Layout to sections folder
