+++ components/elements/index.tsx
@@ -4,7 +4,6 @@ export {
   ProgressCard,
   ShareProgressCard,
 } from './cards';
-export { Layout } from './Layout';
 export { Profile } from './Profile';
 export { SearchBar } from './SearchBar';
 export { ThemeSwitcherButton } from './ThemeSwitcherButton';
