+++ components/sections/index.tsx
@@ -1,2 +1,3 @@
 export { CardGroup } from './CardGroup';
+export { Layout } from './Layout';
 export { ProgressGroup } from './ProgressGroup';
